+++ app/Http/Controllers/SpeakerController.php
@@ -13,7 +13,10 @@ class SpeakerController extends Controller
     public function index()
     {
         $speakers = Speaker::all();
-        return response()->json($speakers);
+        return response()->json([
+            'speakers' => $speakers,
+            'data_count' => $speakers->count() 
+        ], 200);
     }
 
     /**
@@ -29,8 +32,9 @@ class SpeakerController extends Controller
         $speakers->exprestise_areas = $request->exprestise_areas;
 
         return response()->json([
-            "message" => "El ponente ha sido agregado correctamente"
-        ]);
+            "message" => "El ponente ha sido agregado correctamente",
+            'data_count' => 1 
+        ], 201);
     }
 
     /**
@@ -41,12 +45,15 @@ class SpeakerController extends Controller
         $speakers = Speaker::find($id);
 
         if(!empty($speakers)){
-            return response()->json($speakers);
+            return response()->json([
+                'speaker' => $speakers,
+                'data_count' => 1
+            ], 200);
         }
         else{
             return response()->json([
                 "message" => "El ponente no se ha encontrado"
-            ]);
+            ], 404);
         }
     }
 
@@ -57,14 +64,21 @@ class SpeakerController extends Controller
     {
         $speakers = Speaker::find($id);
 
+        if (!$speakers) {
+            return response()->json([
+                'message' => 'El ponente no se ha encontrado'
+            ], 404); 
+        }
+
         $speakers->name = $request->name;
         $speakers->photo_url = $request->photo_url;
         $speakers->social_links = $request->social_links;
         $speakers->exprestise_areas = $request->exprestise_areas;
 
         return response()->json([
-            "message" => "El ponente ha sido actualizado correctamente"
-        ]);
+            "message" => "El ponente ha sido actualizado correctamente",
+            'data_count' => 1
+        ], 200);
     }
 
     /**
@@ -73,10 +87,18 @@ class SpeakerController extends Controller
     public function destroy($id)
     {
         $speakers = Speaker::find($id);
+
+        if (!$speakers) {
+            return response()->json([
+                'message' => 'El ponente no se ha encontrado'
+            ], 404); 
+        }
+
         $speakers->delete();
 
         return response()->json([
-            "message" => "El ponente ha sido borrado correctamente"
-        ]);
+            "message" => "El ponente ha sido borrado correctamente",
+            'data_count' => 0 
+        ], 200);
     }
 }
